requirements.txt
gunicorn==19.7.1
Flask==0.10.1
RPi.GPIO==0.6.3
gevent==1.2.2
